# campus-recruitment-plan/校园招聘接口文档包/OpenAPI接口草案.yaml
openapi: 3.1.0
info:
  title: 校园智能招聘与面试预约平台 API
  version: 1.0.0
  description: 面向学生、企业 HR、管理员的校园招聘与面试预约平台 OpenAPI 草案。
servers:
- url: http://localhost:8080/api
  description: 本地开发环境
- url: https://api.example.com/api
  description: 生产环境占位
...
security:
- bearerAuth: []
paths:
...
  /jobs:
    get:
      tags:
      - Job
      summary: 岗位列表 / 岗位搜索
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseJobPage'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
      parameters:
      - name: keyword
        in: query
        required: false
        description: 关键词
        schema:
          type: string
      - name: city
        in: query
        required: false
        description: 城市
        schema:
          type: string
      - name: salaryMin
        in: query
        required: false
        description: 最低薪资
        schema:
          type: integer
          format: int32
      - name: salaryMax
        in: query
        required: false
        description: 最高薪资
        schema:
          type: integer
          format: int32
      - name: education
        in: query
        required: false
        description: 学历要求
        schema:
          type: string
      - name: experience
        in: query
        required: false
        description: 经验要求
        schema:
          type: string
      - name: sort
        in: query
        required: false
        description: 排序
        schema:
          type: string
          enum:
          - LATEST
          - HOT
      - &id001
        name: pageNum
        in: query
        required: false
        description: 页码
        schema:
          type: integer
          format: int32
          default: 1
      - &id002
        name: pageSize
        in: query
        required: false
        description: 每页数量
        schema:
          type: integer
          format: int32
          default: 10
  /jobs/{id}:
    get:
      tags:
      - Job
      summary: 岗位详情
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseJobDetail'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
      parameters:
      - name: id
        in: path
        required: true
        description: 岗位 ID
        schema:
          type: integer
          format: int64
  /jobs/{id}/favorite:
    post:
      tags:
      - Job
      summary: 收藏岗位
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 岗位 ID
        schema:
          type: integer
          format: int64
    delete:
      tags:
      - Job
      summary: 取消收藏岗位
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 岗位 ID
        schema:
          type: integer
          format: int64
  /company/jobs:
    get:
      tags:
      - CompanyJob
      summary: 企业岗位列表
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseJobPage'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: status
        in: query
        required: false
        description: 岗位状态
        schema:
          type: string
          enum:
          - DRAFT
          - PENDING_REVIEW
          - PUBLISHED
          - REJECTED
          - OFFLINE
          - EXPIRED
      - name: keyword
        in: query
        required: false
        description: 关键词
        schema:
          type: string
      - *id001
      - *id002
    post:
      tags:
      - CompanyJob
      summary: 创建岗位
      responses:
        '201':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseJob'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
  /company/jobs/{id}:
    get:
      tags:
      - CompanyJob
      summary: 企业岗位详情
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseJobDetail'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 岗位 ID
        schema:
          type: integer
          format: int64
    put:
      tags:
      - CompanyJob
      summary: 编辑岗位
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseJob'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 岗位 ID
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobUpdateRequest'
  /company/jobs/{id}/submit:
    put:
      tags:
      - CompanyJob
      summary: 提交岗位审核
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 岗位 ID
        schema:
          type: integer
          format: int64
  /company/jobs/{id}/offline:
    put:
      tags:
      - CompanyJob
      summary: 下架岗位
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 岗位 ID
        schema:
          type: integer
          format: int64
...
  /admin/jobs/{id}:
    get:
      tags:
      - Admin
      summary: 管理员查看岗位详情
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseJobDetail'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 岗位 ID
        schema:
          type: integer
          format: int64
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JobStatus:
      type: string
      enum:
      - DRAFT
      - PENDING_REVIEW
      - PUBLISHED
      - REJECTED
      - OFFLINE
      - EXPIRED
...
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 30002
        message:
          type: string
          example: 重复投递
        data:
          nullable: true
...
    JobCreateRequest:
      type: object
      required:
      - title
      - city
      - salaryUnit
      - description
      - requirement
      properties:
        title:
          type: string
        category:
          type: string
        city:
          type: string
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        salaryUnit:
          type: string
          example: DAY
        education:
          type: string
          example: BACHELOR
        experience:
          type: string
          example: NONE
        description:
          type: string
        requirement:
          type: string
        tags:
          type: array
          items:
            type: string
        expireTime:
          type: string
          format: date-time
    JobUpdateRequest:
      allOf:
      - $ref: '#/components/schemas/JobCreateRequest'
    Job:
      type: object
      properties:
        id:
          type: integer
          format: int64
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
        title:
          type: string
        category:
          type: string
        city:
          type: string
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        salaryUnit:
          type: string
        education:
          type: string
        experience:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        tags:
          type: array
          items:
            type: string
        viewCount:
          type: integer
        applyCount:
          type: integer
        favoriteCount:
          type: integer
        createTime:
          type: string
          format: date-time
    JobDetail:
      allOf:
      - $ref: '#/components/schemas/Job'
      - type: object
        properties:
          description:
            type: string
          requirement:
            type: string
          auditReason:
            type: string
          expireTime:
            type: string
            format: date-time
          favorited:
            type: boolean
          applied:
            type: boolean
...
    JobPage:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/Job'
        total:
          type: integer
          format: int64
          example: 100
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
...
    ApiResponseJob:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/Job'
    ApiResponseJobDetail:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/JobDetail'
    ApiResponseJobPage:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/JobPage'
...